requirements.txt
fastapi==0.109.1
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.0
python-dotenv==1.0.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
passlib
jinja2
pytest
passlib
sentry-sdk[fastapi]
bcrypt
python-jose[cryptography]